各エラーをファイル移動
LexError, LexErrorWithSourceMap を nako_lex_error.js へ移動。
SyntaxErrorWithSourceMap を nako_syntax_error.js へ移動。
