+++ src/nako_syntax_error.js
@@ -18,4 +18,27 @@ class NakoSyntaxError extends Error {
   }
 }
 
-module.exports = NakoSyntaxError
+class NakoSyntaxErrorWithSourceMap extends NakoSyntaxError {
+  /**
+   *@param {import('./nako3').TokenWithSourceMap} token
+   *@param {number} startOffset
+   *@param {number} endOffset
+   *@param {NakoSyntaxError} error
+   */
+  constructor(token, startOffset, endOffset, error) {
+      super(error.msg, error.line, error.fname)
+      /** @readonly */
+      this.token = token
+      /** @readonly */
+      this.startOffset = startOffset
+      /** @readonly */
+      this.endOffset = endOffset
+      /** @readonly */
+      this.error = error
+  }
+}
+
+module.exports = {
+  NakoSyntaxError,
+  NakoSyntaxErrorWithSourceMap,
+}
